# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * خطط سيرك API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: string;
  username: string;
  email: string;
  createdAt: string;
}

export interface RegisterRequest {
  username: string;
  email: string;
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type PlanType = (typeof PlanType)[keyof typeof PlanType];

export const PlanType = {
  رحلة: "رحلة",
  سفر: "سفر",
  طلعة: "طلعة",
  مشوار: "مشوار",
  مخصص: "مخصص",
} as const;

export type PlanStatus = (typeof PlanStatus)[keyof typeof PlanStatus];

export const PlanStatus = {
  خاصة: "خاصة",
  مشتركة: "مشتركة",
  منتهية: "منتهية",
} as const;

export interface Plan {
  id: string;
  title: string;
  description?: string;
  type: PlanType;
  status: PlanStatus;
  hasPassword: boolean;
  shareToken?: string | null;
  notes?: string | null;
  scheduledAt?: string | null;
  userId: string;
  createdAt: string;
  updatedAt: string;
}

export interface Stop {
  id: string;
  planId: string;
  name: string;
  notes?: string | null;
  lat?: number | null;
  lng?: number | null;
  order: number;
  createdAt: string;
}

export type PlanWithStops = Plan & {
  stops: Stop[];
};

export type CreatePlanRequestType =
  (typeof CreatePlanRequestType)[keyof typeof CreatePlanRequestType];

export const CreatePlanRequestType = {
  رحلة: "رحلة",
  سفر: "سفر",
  طلعة: "طلعة",
  مشوار: "مشوار",
  مخصص: "مخصص",
} as const;

export type CreatePlanRequestStatus =
  (typeof CreatePlanRequestStatus)[keyof typeof CreatePlanRequestStatus];

export const CreatePlanRequestStatus = {
  خاصة: "خاصة",
  مشتركة: "مشتركة",
  منتهية: "منتهية",
} as const;

export interface CreatePlanRequest {
  title: string;
  description?: string;
  type: CreatePlanRequestType;
  status: CreatePlanRequestStatus;
  password?: string | null;
  notes?: string | null;
  scheduledAt?: string | null;
}

export type UpdatePlanRequestType =
  (typeof UpdatePlanRequestType)[keyof typeof UpdatePlanRequestType];

export const UpdatePlanRequestType = {
  رحلة: "رحلة",
  سفر: "سفر",
  طلعة: "طلعة",
  مشوار: "مشوار",
  مخصص: "مخصص",
} as const;

export type UpdatePlanRequestStatus =
  (typeof UpdatePlanRequestStatus)[keyof typeof UpdatePlanRequestStatus];

export const UpdatePlanRequestStatus = {
  خاصة: "خاصة",
  مشتركة: "مشتركة",
  منتهية: "منتهية",
} as const;

export interface UpdatePlanRequest {
  title?: string;
  description?: string;
  type?: UpdatePlanRequestType;
  status?: UpdatePlanRequestStatus;
  password?: string | null;
  notes?: string | null;
  scheduledAt?: string | null;
}

export interface CreateStopRequest {
  name: string;
  notes?: string | null;
  lat?: number | null;
  lng?: number | null;
  order: number;
}

export interface UpdateStopRequest {
  name?: string;
  notes?: string | null;
  lat?: number | null;
  lng?: number | null;
  order?: number;
}

export interface ShareResponse {
  shareToken: string;
  shareUrl: string;
}

export interface VerifyPasswordRequest {
  password: string;
}

export interface InviteInfo {
  planId: string;
  title: string;
  description?: string;
  type: string;
  hasPassword: boolean;
  ownerName: string;
}