# lib/api-zod/src/generated/types/createPlanRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * خطط سيرك API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePlanRequestStatus } from "./createPlanRequestStatus";
import type { CreatePlanRequestType } from "./createPlanRequestType";

export interface CreatePlanRequest {
  title: string;
  description?: string;
  type: CreatePlanRequestType;
  status: CreatePlanRequestStatus;
  password?: string | null;
  notes?: string | null;
  scheduledAt?: string | null;
}