# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * خطط سيرك API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanStatus } from "./planStatus";
import type { PlanType } from "./planType";

export interface Plan {
  id: string;
  title: string;
  description?: string;
  type: PlanType;
  status: PlanStatus;
  hasPassword: boolean;
  shareToken?: string | null;
  notes?: string | null;
  scheduledAt?: string | null;
  userId: string;
  createdAt: string;
  updatedAt: string;
}