# lib/api-zod/src/generated/types/shareResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * خطط سيرك API
 * OpenAPI spec version: 0.1.0
 */

export interface ShareResponse {
  shareToken: string;
  shareUrl: string;
}