# lib/api-zod/src/generated/types/stop.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * خطط سيرك API
 * OpenAPI spec version: 0.1.0
 */

export interface Stop {
  id: string;
  planId: string;
  name: string;
  notes?: string | null;
  lat?: number | null;
  lng?: number | null;
  order: number;
  createdAt: string;
}