# lib/api-zod/src/generated/types/updatePlanRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * خطط سيرك API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdatePlanRequestStatus } from "./updatePlanRequestStatus";
import type { UpdatePlanRequestType } from "./updatePlanRequestType";

export interface UpdatePlanRequest {
  title?: string;
  description?: string;
  type?: UpdatePlanRequestType;
  status?: UpdatePlanRequestStatus;
  password?: string | null;
  notes?: string | null;
  scheduledAt?: string | null;
}